Implements authentication context and protected routes
Introduces an authentication context using React Context API to manage user authentication state.

Replaces direct localStorage access for authentication status with the new authentication context. This allows for a more centralized and maintainable approach to authentication.

Updates routing logic to utilize the authentication context and ProtectedRoute component for controlling access to protected routes.

The key now changes based on the current user's ID instead of the local storage which is deprecated.